+++ src/charset.c
@@ -12,6 +12,7 @@
 #include "main.h"
 #include "mbyte.h"
 #include "memline.h"
+#include "memory.h"
 #include "misc1.h"
 #include "misc2.h"
 #include "garray.h"
