+++ src/diff.c
@@ -17,6 +17,7 @@
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "memory.h"
 #include "move.h"
 #include "normal.h"
 #include "option.h"
