+++ src/digraph.c
@@ -12,6 +12,7 @@
 #include "mbyte.h"
 #include "message.h"
 #include "misc2.h"
+#include "memory.h"
 #include "garray.h"
 #include "normal.h"
 #include "screen.h"
