+++ src/ex_docmd.c
@@ -35,6 +35,7 @@
 #include "mark.h"
 #include "mbyte.h"
 #include "memline.h"
+#include "memory.h"
 #include "menu.h"
 #include "message.h"
 #include "misc1.h"
