+++ src/ex_eval.c
@@ -19,6 +19,7 @@
 #include "ex_docmd.h"
 #include "message.h"
 #include "misc2.h"
+#include "memory.h"
 #include "regexp.h"
