+++ src/file_search.c
@@ -50,6 +50,7 @@
 #include "file_search.h"
 #include "charset.h"
 #include "fileio.h"
+#include "memory.h"
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
