+++ src/garray.c
@@ -7,6 +7,7 @@
 #include "vim.h"
 #include "ascii.h"
 #include "misc2.h"
+#include "memory.h"
 #include "path.h"
 #include "garray.h"
