+++ src/hardcopy.c
@@ -24,6 +24,7 @@
 #include "fileio.h"
 #include "mbyte.h"
 #include "memline.h"
+#include "memory.h"
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
