+++ src/hashtab.c
@@ -23,6 +23,7 @@
 #include "vim.h"
 #include "hashtab.h"
 #include "message.h"
+#include "memory.h"
 #include "misc2.h"
 
 // Magic value for algorithm that walks through the array.
