+++ src/mark.c
@@ -24,6 +24,7 @@
 #include "fold.h"
 #include "mbyte.h"
 #include "memline.h"
+#include "memory.h"
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
