+++ src/memfile.c
@@ -41,6 +41,7 @@
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "memory.h"
 #include "os_unix.h"
 #include "path.h"
 #include "ui.h"
