+++ src/path.c
@@ -11,6 +11,7 @@
 #include "garray.h"
 #include "memfile.h"
 #include "memline.h"
+#include "memory.h"
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
