+++ src/regexp.c
@@ -53,6 +53,7 @@
 #include "ex_cmds2.h"
 #include "mark.h"
 #include "memline.h"
+#include "memory.h"
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
