+++ src/spell.c
@@ -314,6 +314,7 @@
 #include "hashtab.h"
 #include "mbyte.h"
 #include "memline.h"
+#include "memory.h"
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
