+++ src/syntax.c
@@ -26,6 +26,7 @@
 #include "indent_c.h"
 #include "mbyte.h"
 #include "memline.h"
+#include "memory.h"
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
