+++ src/undo.c
@@ -94,6 +94,7 @@
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "memory.h"
 #include "crypt.h"
 #include "garray.h"
 #include "option.h"
