+++ src/window.c
@@ -27,6 +27,7 @@
 #include "main.h"
 #include "mark.h"
 #include "memline.h"
+#include "memory.h"
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
